src/main/java/com/example/blog/Hello/HelloController.java
package com.example.blog.Hello;


import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class HelloController {

    @GetMapping("/hello")
    public String hello(){
        return "hello";
    } // hello

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name, @RequestParam("amount") int amount){
        name = "hong";
        amount = 1000;
        return new HelloResponseDto(name, amount);
    } // helloDto


}// HelloController

